Improved alignment-aware typos
Modifications for random_space and skipped_space replacements are now
more robust to contiguous whitespace, single-character words and the
treatment insertions/deletions at the beginning and end of strings,
whether there is untrimmed whitespace or not.

The missing_char modifier from typos may also delete single-character
words, and is replemented here to be alignmnet aware. The deletion of a
single-char word randomly assigns the alignmnet between the adjacent
words, if they exist.